safer pool_free(); don't require matching free_Xbuf type with the original alloc
Use the pool header to store that information.

The corruption checks are still intact.  They weren't using the ph->poolsize
until after those checks anyway, so we don't lose any of that resiliency.

What we gain, though, is not depending on having to match the original
allocation's buffer type when we eventually do a free, which could happen in
very different parts of the code.

pool_free now ignores the passed in 'poolnum' in favor of the poolnum that's
now being stored in the poolheader during the original allocation.

This deprecates all the free_Xbuf() macros (a new, generic, free_buf() has been
created), but they're left there to keep compatability with the swarms of them
in use.  We may remove them later and do a massive search/replace, but for now,
let's see how this goes.

(This has been running clean under valgrind for me for days)